split up reducers
